Improve test coverage for flask filters
Tests when filters are no-ops (flask context is inactive)
